layers: Move layout implementation
