+++ layers/pipeline_layout_state.h
@@ -3,6 +3,8 @@
 #include <vector>
 #include <memory>
 #include "base_node.h"
+#include "hash_util.h"
+#include "hash_vk_types.h"
 
 // Fwd declarations -- including descriptor_set.h creates an ugly include loop
 namespace cvdescriptorset {
